@@ -2,7 +2,7 @@
   "name": "server",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "index.ts",
   "scripts": {
     "start-dev": "nodemon index.ts",
     "start": "node index.ts"
@@ -18,21 +18,21 @@
     "cloudinary": "^1.32.0",
     "cors": "^2.8.5",
     "dotenv": "^16.0.3",
-    "eslint": "^8.28.0",
     "express": "^4.18.2",
     "express-graphql": "^0.12.0",
     "graphql": "^15.8.0",
-    "husky": "^8.0.2",
     "jsonwebtoken": "^9.0.0",
     "mongoose": "^6.7.5",
     "nodemailer": "^6.9.0",
-    "prettier": "^2.8.0",
     "serverless": "^3.26.0",
     "ts-node": "^10.9.1",
-    "typescript": "^4.9.4",
-    "zod": "^3.20.2"
+    "typescript": "^4.9.4"
   },
   "devDependencies": {
-    "nodemon": "^2.0.20"
+    "eslint": "^8.32.0",
+    "husky": "^8.0.3",
+    "nodemon": "^2.0.20",
+    "prettier": "^2.8.3",
+    "zod": "^3.20.2"
   }
 }
